Calculate starting trunk index
When the collector is restarted, it needs to pick up from where
it left off and continue collecting, instead of always starting
from index 0 and truncating existing storage.

The heuristic is to look at the timestamp of all qualified
trunk files and start from the subsequent index of the newest.